Convert to half/float16
Looks good but constantK is broken and
still some streaks

+++ ntsc-mtl-demo/Filters/MetalKernels/IIRFilterSample.metal
@@ -10,13 +10,13 @@ using namespace metal;
 
 kernel void iirFilterSample
 (
- texture2d<float, access::read> inputTexture [[texture(0)]],
- texture2d<float, access::read> zTex0 [[texture(1)]],
- texture2d<float, access::read_write> filteredSampleTexture [[texture(2)]],
- constant float &num0 [[buffer(0)]],
+ texture2d<half, access::read> inputTexture [[texture(0)]],
+ texture2d<half, access::read> zTex0 [[texture(1)]],
+ texture2d<half, access::read_write> filteredSampleTexture [[texture(2)]],
+ constant half &num0 [[buffer(0)]],
  uint2 gid [[thread_position_in_grid]]
  ) {
-    float4 result = zTex0.read(gid) + (num0 * inputTexture.read(gid));
+    half4 result = zTex0.read(gid) + (num0 * inputTexture.read(gid));
     result.w = 1.0;
     filteredSampleTexture.write(result, gid);
 }
